move execute to checklist and remove get_streaming_checklist
